V0.0.3.4
Added in more visual indicators for the "file Selection and Deletion page"

The schedule now will run each day if "Daily" is set as true if the program is open

Disabled "copy selected" button when using the individual copy

The overwrite button will now change between "Copy" and "Overwrite" depending if its checked or not.

Add a button for the confirmation on overwrite

Redid some of the system. Instead of taking from the file name used in the replace path, dump it in the targeted directory and use the name from the copy location. (Help clear up confusion)

Add a feature that tell you the latest date in the "File Selection and Deletion" (prompt)